move testing utilities to their own modules
